+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.3.2] - 2025-02-12
 ### Changed
 - Now prints the user agent to the log.
 - Set the certificate serial number to a random number. The number should not be absolutely relied upon to be unique.
@@ -85,6 +87,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Initial release
 
+[0.3.2]: https://github.com/isambard-sc/conch/releases/tag/0.3.2
 [0.3.1]: https://github.com/isambard-sc/conch/releases/tag/0.3.1
 [0.3.0]: https://github.com/isambard-sc/conch/releases/tag/0.3.0
 [0.2.0]: https://github.com/isambard-sc/conch/releases/tag/0.2.0